actually be able to connect to the ws server

+++ taiko-rs-common/src/tables/users_table.rs
@@ -4,7 +4,7 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "users")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub user_id: i64,
+    pub user_id: i32,
     pub username: String,
     pub password: String,
     pub email: String,
